+++ aasrp/static/aasrp/javascript/aa-srp.js
@@ -62,17 +62,3 @@ if (typeof aaSrpJsSettingsOverride !== 'undefined') {
         aaSrpJsSettingsOverride
     );
 }
-
-/**
- * Generates a copy-to-clipboard icon with the specified data and title.
- *
- * @param {string} data - The data to be copied to the clipboard.
- * @param {string} title - The title for the tooltip.
- * @returns {`<span class="copy-to-clipboard-icon">${string}</span>`} The HTML string for the copy-to-clipboard-icon icon.
- */
-const copyToClipboardIcon = (data, title) => { // eslint-disable-line no-unused-vars
-    const ctcIconClasses = 'copy-to-clipboard fa-regular fa-copy ms-2 cursor-pointer';
-    const ctcIcon = `<i class="${ctcIconClasses}" data-clipboard-text="${data}" data-bs-tooltip="aa-srp" aria-label="${title}" title="${title}"></i>`;
-
-    return `<span class="copy-to-clipboard-icon">${ctcIcon}</span>`;
-};